Add logging.

+++ app/src/main/java/net/sourceforge/opencamera/preview/Preview.java
@@ -1012,6 +1012,8 @@ public class Preview implements SurfaceHolder.Callback, TextureView.SurfaceTextu
             // if nextVideoFileInfo is not-null, it means we received MEDIA_RECORDER_INFO_MAX_FILESIZE_APPROACHING but not
             // MEDIA_RECORDER_INFO_NEXT_OUTPUT_FILE_STARTED, so it is the application responsibility to create the zero-size
             // video file that will have been created
+            if( MyDebug.LOG )
+                Log.d(TAG, "delete ununused next video file");
             nextVideoFileInfo.close();
             applicationInterface.deleteUnusedVideo(nextVideoFileInfo.video_method, nextVideoFileInfo.video_uri, nextVideoFileInfo.video_filename);
         }
